Refactor: 옷 추천 조회 중복 메서드 extract (#52)
